fixes test activity
