+++ services/web/server/tests/unit/isolated/test_rest.py
@@ -9,7 +9,6 @@ from unittest.mock import MagicMock
 import pytest
 from aiohttp import web
 from aiohttp.test_utils import TestClient
-from models_library.users import UserID
 from pytest_mock import MockerFixture
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from pytest_simcore.helpers.utils_assert import assert_status
@@ -74,31 +73,6 @@ async def test_frontend_config(
 
 
 @pytest.fixture
-def mock_login_required(mocker: MockerFixture) -> UserID:
-    user_id = 1
-    # patches @login_required decorator
-    # avoids having to start database etc...
-    mocker.patch(
-        "simcore_service_webserver.login.decorators.check_user_authorized",
-        spec=True,
-        return_value=user_id,
-    )
-
-    mocker.patch(
-        "simcore_service_webserver.login.decorators.check_user_permission",
-        spec=True,
-        return_value=None,
-    )
-
-    mocker.patch(
-        "simcore_service_webserver.login.decorators.get_product_name",
-        spec=True,
-        return_value="osparc",
-    )
-    return user_id
-
-
-@pytest.fixture
 async def mock_redis_client(
     client: TestClient,
     mocker: MockerFixture,
@@ -139,7 +113,7 @@ async def test_get_scheduled_maintenance(
     api_version_prefix: str,
     redis_maintenance_data: dict[str, str],
     expected: type[web.HTTPException],
-    mock_login_required: UserID,
+    mocked_login_required: None,
     mock_redis_client: MagicMock,
 ):
     assert client.app
